Added some market interactions (WIP)

+++ Assets/Scripts/Infrastructure/Core/Ship/ShipModel.cs
@@ -11,20 +11,20 @@ namespace Infrastructure.Core.Ship
         public int currentShieldAmount;
         public float currentEnergyAmount;
 
-        public Dictionary<ShipStats, float> shipStats;
+        public Dictionary<ShipStats, int> shipStats;
 
         protected Dictionary<ShipParts, ShipPart> shipParts;
 
         public ShipModel()
         {
             shipParts = new Dictionary<ShipParts, ShipPart>();
-            shipStats = new Dictionary<ShipStats, float>();
+            shipStats = new Dictionary<ShipStats, int>();
         }
 
         public void AddPart(ShipParts partName, ShipPart part)
         {
             shipParts.Add(partName, part);
-            foreach(KeyValuePair<ShipStats, float> stat in part.stats)
+            foreach(KeyValuePair<ShipStats, int> stat in part.stats)
             {
                 shipStats.Add(stat.Key, stat.Value);
             }
